requirements.txt
flask
pycryptodome
uwsgi